Fix OLED preview sizing/race, dedupe LED glow code, resolve /screens card previews
services/displays/preview_render.py:
- render_oled_elements_preview() now reads the operator's actual panel size
  from HardwareSettings instead of hardcoding 128x64, falling back to that
  default only when no DB/app context is available.
- Replace the unittest.mock.patch on shared app_core.oled module state with
  a controller built via __new__ + a _NullOLEDDevice stand-in, so concurrent
  preview requests under gunicorn's gevent workers can't race on a global
  patch.
- Dedupe the LED glow-rendering loop shared by render_led_preview() and
  render_led_elements_preview() into one _render_led_glow() helper.
- Split the file (506 lines, over the repo's 350-line guideline) into
  scripts/{preview_render_common,dotmatrix_preview_font,led_preview_render,
  vfd_preview_render,oled_preview_render}.py, with preview_render.py left as
  a thin re-export facade. These live under scripts/ rather than
  services/displays/ deliberately: tests/test_display_preview_render.py
  loads preview_render.py directly by path to avoid pulling in
  services.displays's Flask-heavy __init__, which sibling modules inside
  that package would defeat.

templates/screens.html:
- The /screens list page was rendering raw, unresolved "{status.status}"-
  style template placeholders on every card instead of realistic sample
  values. Add hydrateScreenPreviews(), which fetches the existing
  GET /api/screens/<id>/preview endpoint (already used by the screen
  editor's live preview) for each screen with bounded concurrency, and
  prefer that resolved data in buildScreenPreview() once it lands.
